lift checkpoint ordering restriction for efo
